refactor(releasechangelog): refactor the name of release changelog interface
