+++ packages/sfpowerscripts-cli/package.json
@@ -1,7 +1,7 @@
 {
     "name": "@dxatscale/sfpowerscripts",
     "description": "A build system for package based development in Salesforce,delivered as a sfdx plugin that can be implemented in any CI/CD system of choice.Read more about the plugin and details here - https://dxatscale.gitbook.io/sfpowerscripts/",
-    "version": "19.4.9",
+    "version": "19.4.10",
     "author": "dxatscale",
     "release": "September 22",
     "bin": {
