+++ packages/sfpowerscripts-cli/tests/impl/changelog/CommitUpdater.test.ts
@@ -2,7 +2,7 @@ import { expect } from '@jest/globals';
 import CommitUpdater from '../../../src/impl/changelog/CommitUpdater';
 import ReadPackageChangelog from '../../../src/impl/changelog/ReadPackageChangelog';
 import { Changelog as PackageChangelog } from '@dxatscale/sfpowerscripts.core/lib/changelog/interfaces/GenericChangelogInterfaces';
-import { Release } from '../../../src/impl/changelog/ReleaseChangelogInterfaces';
+import { Release } from '../../../src/impl/changelog/ReleaseChangelog';
 const path = require('path');
 import * as fs from 'fs-extra';
